Refactor the dpos code to organize it better
